+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -4,7 +4,7 @@ Release History
 ===============
 unreleased
 +++++++++++++++++++
-* `vm/vmss create`: don't throw when failed to retrieve the image to extract out the plan information 
+* `vm/vmss create`: fix issue where the command would throw an error if unable to extract plan information from an image. 
 * `vmss create`: fix a crash when create a scaleset with an internal LB
 * `vm availability-set create`: Fix issue where --no-wait argument did not work.